DF-548:relocate data class definitions
